pyproject.toml
[project]
name = "framepack"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "~=3.10"

dependencies = [
    "accelerate==1.6.0",
    "av==12.1.0",
    "decord>=0.6.0",
    "diffusers==0.33.1",
    "einops>=0.8.1",
    "gradio>=5.25.2",
    "jinja2>=3.1.6",
    "numpy==1.26.2",
    "opencv-contrib-python>=4.11.0.86",
    "peft>=0.15.2",
    "pillow==11.1.0",
    "python-dotenv>=1.1.0",
    "python-magic>=0.4.27",
    "requests>=2.32.3",
    "safetensors>=0.5.3",
    "scipy==1.12.0",
    "sentencepiece==0.2.0",
    "setuptools>=78.1.0",
    "tenacity>=9.1.2",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchsde==0.2.6",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "transformers==4.46.2",
    "truststore>=0.10.1",
]

[project.optional-dependencies]
sage = [
    "sageattention==1.0.6",
]
#sage2 = [
#    "sageattention",
#]
flash = [
    "flash-attn>=2.7.4.post1",
]
xformers = [
    "xformers>=0.0.29.post3",
]
runpod = [
    "runpod>=1.7.9",
]

[tool.uv]
no-build-isolation-package = [
    "sageattention",
    "xformers",
    "flash-attn",
]

#conflicts = [
#    [
#      { extra = "sage" },
#      { extra = "sage2" },
#    ],
#]

[tool.uv.sources]
#sageattention = [
#    { git = "https://github.com/thu-ml/SageAttention", extra = "sage2" },
#]